Restructure TLS (#274)
The plan consists of restructuring TLS following best practices to simplify the understanding of each component.
