+++ huginn-net-tls/src/analyzer.rs
@@ -1,0 +1,341 @@
+use crate::error::HuginnNetTlsError;
+use crate::filter::raw as raw_filter;
+use crate::filter::FilterConfig;
+use crate::output::TlsClientOutput;
+use crate::parser::packet::{parse_packet, IpPacket};
+use crate::parser::TlsClientHelloReader;
+use crate::process::parallel::WorkerPool;
+use crate::process::{process_ipv4_packet, process_ipv6_packet, FlowKey};
+use pcap_file::pcap::PcapReader;
+use pnet::datalink::{self, Channel, Config};
+use std::fs::File;
+use std::sync::atomic::{AtomicBool, Ordering};
+use std::sync::mpsc::Sender;
+use std::sync::Arc;
+use tracing::{debug, error};
+use ttl_cache::TtlCache;
+
+/// Configuration for parallel processing
+///
+/// Controls the behavior of worker threads in parallel mode.
+#[derive(Debug, Clone)]
+struct ParallelConfig {
+    num_workers: usize,
+    queue_size: usize,
+    batch_size: usize,
+    timeout_ms: u64,
+}
+
+/// A TLS-focused passive fingerprinting analyzer using JA4 methodology.
+///
+/// Extracts JA4 client-hello fingerprints from raw network traffic without
+/// decrypting TLS, no database or external matcher is required.
+///
+/// # Examples
+///
+/// **Sequential: JA4 fingerprints from a PCAP file:**
+///
+/// ```no_run
+/// use huginn_net_tls::HuginnNetTls;
+/// use std::sync::mpsc;
+///
+/// let (tx, rx) = mpsc::channel();
+/// HuginnNetTls::new(1000).analyze_pcap("capture.pcap", tx, None).unwrap();
+/// for output in rx {
+///     println!("{output}");
+/// }
+/// ```
+///
+/// **Sequential: live capture on a network interface:**
+///
+/// ```no_run
+/// use huginn_net_tls::HuginnNetTls;
+/// use std::sync::mpsc;
+///
+/// let (tx, rx) = mpsc::channel();
+/// HuginnNetTls::new(10000).analyze_network("eth0", tx, None).unwrap();
+/// ```
+///
+/// **Parallel: worker pool for high-throughput live capture:**
+///
+/// ```no_run
+/// use huginn_net_tls::HuginnNetTls;
+/// use std::sync::mpsc;
+///
+/// let (tx, rx) = mpsc::channel();
+/// let mut analyzer = HuginnNetTls::new(10000).with_parallel(4, 100, 32, 10);
+/// analyzer.init_pool(tx.clone()).unwrap();
+/// analyzer.analyze_network("eth0", tx, None).unwrap();
+/// ```
+pub struct HuginnNetTls {
+    tcp_flows: TtlCache<FlowKey, TlsClientHelloReader>,
+    parallel_config: Option<ParallelConfig>,
+    worker_pool: Option<Arc<WorkerPool>>,
+    filter_config: Option<FilterConfig>,
+    max_connections: usize,
+}
+
+impl HuginnNetTls {
+    /// Creates a new instance of `HuginnNetTls` in sequential mode (single-threaded).
+    ///
+    /// # Parameters
+    /// - `max_connections`: Maximum number of TCP flows to track
+    pub fn new(max_connections: usize) -> Self {
+        Self {
+            tcp_flows: TtlCache::new(max_connections),
+            parallel_config: None,
+            worker_pool: None,
+            filter_config: None,
+            max_connections,
+        }
+    }
+
+    /// Enable parallel processing (builder pattern).
+    ///
+    /// # Parameters
+    /// - `num_workers`: Number of worker threads (recommended: 2-4 on 8-core systems)
+    /// - `queue_size`: Size of packet queue per worker (typical: 100-200)
+    /// - `batch_size`: Maximum packets to process in one batch (typical: 16-64, recommended: 32)
+    /// - `timeout_ms`: Worker receive timeout in milliseconds (typical: 5-50, recommended: 10)
+    pub fn with_parallel(
+        mut self,
+        num_workers: usize,
+        queue_size: usize,
+        batch_size: usize,
+        timeout_ms: u64,
+    ) -> Self {
+        self.parallel_config =
+            Some(ParallelConfig { num_workers, queue_size, batch_size, timeout_ms });
+        self
+    }
+
+    /// Configure packet filtering (builder pattern).
+    pub fn with_filter(mut self, config: FilterConfig) -> Self {
+        self.filter_config = Some(config);
+        self
+    }
+}
+
+impl HuginnNetTls {
+    /// Get worker pool statistics (only available in parallel mode, after analyze_* is called)
+    pub fn stats(&self) -> Option<crate::process::parallel::PoolStats> {
+        self.worker_pool.as_ref().map(|pool| pool.stats())
+    }
+
+    /// Get a reference to the worker pool (only available in parallel mode, after analyze_* is called)
+    pub fn worker_pool(&self) -> Option<Arc<WorkerPool>> {
+        self.worker_pool.clone()
+    }
+
+    /// Initialize the worker pool (only for parallel mode, called automatically by analyze_*)
+    ///
+    /// # Errors
+    /// - If the worker pool creation fails.
+    pub fn init_pool(&mut self, sender: Sender<TlsClientOutput>) -> Result<(), HuginnNetTlsError> {
+        let config = self.parallel_config.as_ref().ok_or_else(|| {
+            HuginnNetTlsError::Parse(
+                "Parallel config not set. Use with_parallel() to enable parallel processing"
+                    .to_string(),
+            )
+        })?;
+        if self.worker_pool.is_none() {
+            let worker_pool = Arc::new(WorkerPool::new(
+                config.num_workers,
+                config.queue_size,
+                config.batch_size,
+                config.timeout_ms,
+                sender,
+                self.max_connections,
+                self.filter_config.clone(),
+            )?);
+            self.worker_pool = Some(worker_pool);
+        }
+        Ok(())
+    }
+
+    fn process_with<F>(
+        &mut self,
+        packet_fn: F,
+        sender: Sender<TlsClientOutput>,
+        cancel_signal: Option<Arc<AtomicBool>>,
+    ) -> Result<(), HuginnNetTlsError>
+    where
+        F: FnMut() -> Option<Result<Vec<u8>, HuginnNetTlsError>>,
+    {
+        if self.parallel_config.is_some() {
+            self.process_parallel(packet_fn, cancel_signal)
+        } else {
+            self.process_sequential(packet_fn, sender, cancel_signal)
+        }
+    }
+
+    fn process_parallel<F>(
+        &mut self,
+        mut packet_fn: F,
+        cancel_signal: Option<Arc<AtomicBool>>,
+    ) -> Result<(), HuginnNetTlsError>
+    where
+        F: FnMut() -> Option<Result<Vec<u8>, HuginnNetTlsError>>,
+    {
+        let worker_pool = self
+            .worker_pool
+            .as_ref()
+            .ok_or_else(|| {
+                HuginnNetTlsError::Parse(
+                    "Worker pool not initialized. Call init_pool() before processing".to_string(),
+                )
+            })?
+            .clone();
+
+        while let Some(packet_result) = packet_fn() {
+            if let Some(ref cancel) = cancel_signal {
+                if cancel.load(Ordering::Relaxed) {
+                    debug!("Cancellation signal received, stopping packet processing");
+                    break;
+                }
+            }
+
+            match packet_result {
+                Ok(packet) => {
+                    let _ = worker_pool.dispatch(packet);
+                }
+                Err(e) => {
+                    error!("Failed to read packet: {e}");
+                }
+            }
+        }
+
+        worker_pool.shutdown();
+        Ok(())
+    }
+
+    fn process_sequential<F>(
+        &mut self,
+        mut packet_fn: F,
+        sender: Sender<TlsClientOutput>,
+        cancel_signal: Option<Arc<AtomicBool>>,
+    ) -> Result<(), HuginnNetTlsError>
+    where
+        F: FnMut() -> Option<Result<Vec<u8>, HuginnNetTlsError>>,
+    {
+        while let Some(packet_result) = packet_fn() {
+            if let Some(ref cancel) = cancel_signal {
+                if cancel.load(Ordering::Relaxed) {
+                    debug!("Cancellation signal received, stopping packet processing");
+                    break;
+                }
+            }
+
+            match packet_result {
+                Ok(packet) => match self.process_packet(&packet) {
+                    Ok(Some(result)) => {
+                        if sender.send(result).is_err() {
+                            error!("Receiver dropped, stopping packet processing");
+                            break;
+                        }
+                    }
+                    Ok(None) => {
+                        debug!("No TLS found, continuing packet processing");
+                    }
+                    Err(tls_error) => {
+                        debug!("Error processing packet: {tls_error}");
+                    }
+                },
+                Err(e) => {
+                    error!("Failed to read packet: {e}");
+                }
+            }
+        }
+        Ok(())
+    }
+
+    /// Captures and analyzes packets on the specified network interface.
+    ///
+    /// # Errors
+    /// - If the network interface cannot be found or a channel cannot be created.
+    pub fn analyze_network(
+        &mut self,
+        interface_name: &str,
+        sender: Sender<TlsClientOutput>,
+        cancel_signal: Option<Arc<AtomicBool>>,
+    ) -> Result<(), HuginnNetTlsError> {
+        let interfaces = datalink::interfaces();
+        let interface = interfaces
+            .into_iter()
+            .find(|iface| iface.name == interface_name)
+            .ok_or_else(|| {
+                HuginnNetTlsError::Parse(format!(
+                    "Could not find network interface: {interface_name}"
+                ))
+            })?;
+
+        debug!("Using network interface: {}", interface.name);
+
+        let config = Config { promiscuous: true, ..Config::default() };
+
+        let (_tx, mut rx) = match datalink::channel(&interface, config) {
+            Ok(Channel::Ethernet(tx, rx)) => (tx, rx),
+            Ok(_) => return Err(HuginnNetTlsError::Parse("Unhandled channel type".to_string())),
+            Err(e) => {
+                return Err(HuginnNetTlsError::Parse(format!("Unable to create channel: {e}")))
+            }
+        };
+
+        self.process_with(
+            move || match rx.next() {
+                Ok(packet) => Some(Ok(packet.to_vec())),
+                Err(e) => {
+                    Some(Err(HuginnNetTlsError::Parse(format!("Error receiving packet: {e}"))))
+                }
+            },
+            sender,
+            cancel_signal,
+        )
+    }
+
+    /// Analyzes packets from a PCAP file.
+    ///
+    /// # Errors
+    /// - If the PCAP file cannot be opened or read.
+    pub fn analyze_pcap(
+        &mut self,
+        pcap_path: &str,
+        sender: Sender<TlsClientOutput>,
+        cancel_signal: Option<Arc<AtomicBool>>,
+    ) -> Result<(), HuginnNetTlsError> {
+        let file = File::open(pcap_path)
+            .map_err(|e| HuginnNetTlsError::Parse(format!("Failed to open PCAP file: {e}")))?;
+        let mut pcap_reader = PcapReader::new(file)
+            .map_err(|e| HuginnNetTlsError::Parse(format!("Failed to create PCAP reader: {e}")))?;
+
+        self.process_with(
+            move || match pcap_reader.next_packet() {
+                Some(Ok(packet)) => Some(Ok(packet.data.to_vec())),
+                Some(Err(e)) => {
+                    Some(Err(HuginnNetTlsError::Parse(format!("Error reading PCAP packet: {e}"))))
+                }
+                None => None,
+            },
+            sender,
+            cancel_signal,
+        )
+    }
+
+    fn process_packet(
+        &mut self,
+        packet: &[u8],
+    ) -> Result<Option<TlsClientOutput>, HuginnNetTlsError> {
+        if let Some(ref filter) = self.filter_config {
+            if !raw_filter::apply(packet, filter) {
+                debug!("Filtered out packet before parsing");
+                return Ok(None);
+            }
+        }
+
+        match parse_packet(packet) {
+            IpPacket::Ipv4(ipv4) => process_ipv4_packet(&ipv4, &mut self.tcp_flows),
+            IpPacket::Ipv6(ipv6) => process_ipv6_packet(&ipv6, &mut self.tcp_flows),
+            IpPacket::None => Ok(None),
+        }
+    }
+}
